ARROW-6484: [Java] Enable create indexType for DictionaryEncoding according to dictionary value count
Related to [ARROW-6484](https://issues.apache.org/jira/browse/ARROW-6484).

Currently, when create DictionaryEncoding, we need to specify indexType, and it use Int(32, true) as default if this value is null.
Actually, when dictionary valueCount is small, we should use Int(8,true)/Int(16,true) instead to reduce memory allocation.
This issue is about to provide API for creating indexType according to valueCount and apply it to avro adapter for enum type.

Closes #5321 from tianchen92/ARROW-6484 and squashes the following commits:

fc1931b4b <tianchen>  Enable create indexType for DictionaryEncoding according to dictionary value count

Authored-by: tianchen <niki.lj@alibaba-inc.com>

+++ java/adapter/avro/src/test/java/org/apache/arrow/AvroToArrowTest.java
@@ -131,7 +131,8 @@ public class AvroToArrowTest extends AvroTestBase {
         new GenericData.EnumSymbol(schema, "DIAMONDS"),
         new GenericData.EnumSymbol(schema, "CLUBS"),
         new GenericData.EnumSymbol(schema, "SPADES"));
-    List<Integer> expectedIndices = Arrays.asList(0, 1, 2, 3, 0);
+
+    List<Byte> expectedIndices = Arrays.asList((byte)0, (byte)1, (byte)2, (byte)3, (byte)0);
 
     VectorSchemaRoot root = writeAndRead(schema, data);
     FieldVector vector = root.getFieldVectors().get(0);
